+++ support/ebpf/types.h
@@ -411,7 +411,7 @@ typedef struct PyProcInfo {
   u64 autoTLSKeyAddr;
   u64 noneStructAddr;
   u16 version;
-  s8 tls_offset;
+  s16 tls_offset;
   TSDInfo tsdInfo;
   // The Python object member offsets
   u8 PyThreadState_frame;
